+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # clever-tools changelog
 
-## Unrelease (????-??-??)
+## Unreleased (????-??-??)
+
+* add config file and auth source to `clever diag`
 
 ## 2.11.0 (2023-07-25)